añadido los mensajes (session start) para los modulos

+++ Crud/clientes/eliminar_cliente.php
@@ -1,17 +1,19 @@
 <?php
-include $_SERVER['DOCUMENT_ROOT'] . '/TravelEase/includes/header.php';
+session_start();
 include $_SERVER['DOCUMENT_ROOT'] . '/TravelEase/includes/conexion.php';
-$id_cliente = $_GET['id'];
 
-// Eliminar el cliente
-$sql = "DELETE FROM Clientes WHERE id_cliente = $id_cliente";
-if ($conn->query($sql) === TRUE) {
-    echo "<div class='alert alert-success'>Cliente eliminado con éxito</div>";
-} else {
-    echo "<div class='alert alert-danger'>Error: " . $conn->error . "</div>";
+if ($_SERVER['REQUEST_METHOD'] == 'POST') {
+    $id_cliente = $_POST['id_cliente'];
+
+    $sql = "DELETE FROM Clientes WHERE id_cliente = $id_cliente";
+    if ($conn->query($sql) === TRUE) {
+        $_SESSION['success'] = "cliente eliminado con éxito.";
+        header("Location: /TravelEase/crud/clientes/clientes.php");
+        exit();
+    } else {
+        $_SESSION['error'] = "Error: " . $conn->error;
+        header("Location: /TravelEase/crud/clientes/clientes.php");
+        exit();
+    }
 }
-
-echo "<a href='clientes.php' class='btn btn-primary'>Volver a la lista de clientes</a>";
 ?>
-
-<?php include $_SERVER['DOCUMENT_ROOT'] . '/TravelEase/includes/footer.php'; ?>
