+++ Crud/reservas/eliminar_reserva.php
@@ -1,14 +1,19 @@
 <?php
-include $_SERVER['DOCUMENT_ROOT'] . '/TravelEase/includes/header.php';
+session_start();
 include $_SERVER['DOCUMENT_ROOT'] . '/TravelEase/includes/conexion.php';
 
-$id_reserva = $_GET['id'];
+if ($_SERVER['REQUEST_METHOD'] == 'POST') {
+    $id_reserva = $_POST['id_reserva'];
 
-$sql = "DELETE FROM Reservas WHERE id_reserva = $id_reserva";
-
-if ($conn->query($sql) === TRUE) {
-    header('Location: reservas.php?msg=Reserva eliminada con éxito');
-} else {
-    echo "Error al eliminar reserva: " . $conn->error;
+    $sql = "DELETE FROM Reservas WHERE id_reserva = $id_reserva";
+    if ($conn->query($sql) === TRUE) {
+        $_SESSION['success'] = "Reserva eliminada con exito.";
+        header("Location: /TravelEase/crud/reservas/reservas.php");
+        exit();
+    } else {
+        $_SESSION['error'] = "Error: " . $conn->error;
+        header("Location: /TravelEase/crud/reservas/reservas.php");
+        exit();
+    }
 }
 ?>
