+++ Crud/transportes/eliminar_transporte.php
@@ -1,17 +1,19 @@
 <?php
-include $_SERVER['DOCUMENT_ROOT'] . '/TravelEase/includes/header.php';
+session_start();
 include $_SERVER['DOCUMENT_ROOT'] . '/TravelEase/includes/conexion.php';
-$id_transporte = $_GET['id'];
 
-// Eliminar el transporte
-$sql = "DELETE FROM Transportes WHERE id_transporte = $id_transporte";
-if ($conn->query($sql) === TRUE) {
-    echo "<div class='alert alert-success'>Transporte eliminado con éxito</div>";
-} else {
-    echo "<div class='alert alert-danger'>Error: " . $conn->error . "</div>";
+if ($_SERVER['REQUEST_METHOD'] == 'POST') {
+    $id_transporte = $_POST['id_transporte'];
+
+    $sql = "DELETE FROM Transportes WHERE id_transporte = $id_transporte";
+    if ($conn->query($sql) === TRUE) {
+        $_SESSION['success'] = "Transporte actualizado con éxito.";
+        header("Location: /TravelEase/crud/transportes/transportes.php");
+        exit();
+    } else {
+        $_SESSION['error'] = "Error: " . $conn->error;
+        header("Location: /TravelEase/crud/transportes/transportes.php");
+        exit();
+    }
 }
-
-echo "<a href='transportes.php' class='btn btn-primary'>Volver a la lista de transportes</a>";
 ?>
-
-<?php include $_SERVER['DOCUMENT_ROOT'] . '/TravelEase/includes/footer.php'; ?>
